Detection of imcomplete types
In header files argument or return types are often forward-declared
to reduce unecessary compilation dependencies. Our meta-data however
depends on the complete type if the variant must hold a value of this
type.
For this reason the parser emits a warning and doesn't generate the
metadata for the declarations that depend on imcomplete types.
I still have to come up with a way for the programmer to provide
the necessary header files manually.

Note: references and pointers to forward-declared classes are _not_
considered to be incomplete.