Reduced meta-data bloat
- The technique of type erasure generates considerable template code bloat.
I've reduced some of this bloat by refactoring methods that didn't need
to be virtual thereby reducing the vtable and the number of methos generated.
In some cases its possible to eliminate the type hierarchy completely.

- As the rrti type information also uses space, i've added support to
disable every usage that the library does of typeid or typeinfo.
This can be done by adding -DNO_RTTI to the compiler command line
Whit some compilers the flag -fno-rtti can be specified in addition.
With -DNO_RTTI the code makes no usage of dynamic_cast or typeid
but it still uses the exception handling cast.
With clang++ the exception cast work even with -fno-rtti and
g++ according to the documentation apparently does too.