Dynamic proxies!
Basically the user takes a bunch of classes that qualify as interfaces
and creates a proxy object. For each method of the interfaces he can
add an implementation using functions, C++11 lambdas or anything else that
has a matching operator().
From the lua API the user can use lua closures to implement individual methods.

The corresponding parser functionality is yet to be implemented.