The parser now generates stubs for interfaces
A class is considered an interface if it follows certain rules.
The rules are listed in definitions.h. For now some rules are a little
too severe, but if experience shows that there is no problem we can relax
them.

The method Class.isInterface returns true for classes that have
a registered proxy stub and can be used to instantiate dynamic proxies