Better proxy implementation
The problem with the old implementation was that you had to keep
the Proxy handle alive during the entire life of the reference
to the interface class. Now you can use the Proxy object only to
configure a dynamic proxy and then discard it. When the VariantValue
containing the reference to the stub dies, all other internal structures
are release